Add files via upload
an engineer is interested in the effects of cutting speed (A), tool geometry (B), and cutting angle (C) on the life (in hours) of a machine tool. Two levels of each factor are chosen and three replicates of a 2^3 factorial design are run. Residual analysis will be carried out - fulfilling the normality assumption